refactor: 도메인 디렉토리 위치 수정
